TT-1601 Integration tests ensure that the MSA can generate test tool json
Should ensure that the json interface the test tool uses stays in sync
with the matching service adapter.
Tests that the json used in each of the default tests in the resources
folder of the test tool can be produced by the matching service adapter.
N.B. Does not ensure that the schema which validates user supplied test
cases remains up to date.

Authors: @rachaelbooth