EID-467: Re-factor MSA validation logic
This is a bit of a monster of a commit so a brief explanation about
it's origins is below.

Before this change, an incoming AQR would follow one of 2 flows.
1. AQR assertion from country -> eIDAS flow
2. AQR assertion from IDP -> Verify flow

These two flows were implemented very differently with eIDAS using
verify-validation and Verify using the transformers.
When adding user account creation (UAC), we had difficulty fitting
the validation/unwrapping of the AQR into either flow when trying
to implement it for the eIDAS flow.
We decided to bring the two flows of control closer in line with each
other by making them more procedural and more similar to the VSP.
Validation and translation are split into seperate steps and as much
shared behaviour is included in the abstract base classes. The outbound
transformers remain the same and the outputs of the translation steps
were shimmed to create the required POJO to generate these Responses.

UAC now uses almost entirely the same steps as matching with the
exception of having to extract the requested attributes. This makes
for clearer, more maintainable code as the actual assertion is identical
between matching and UAC (due to the end-to-end trust we maintain)

When trying to read this change I would suggest:
1. Look at the diff in src/integration test to confirm that only some
minor error messages were updated
2. Look into the matching and UAC resource to see the large scale
logical steps being taken
3. Look into the AQR service to see the outer object validation and
logic for which flow to use
4. Look at the abstract Assertion service and the 2 concrete classes
to see the shared behaviour/differences between the 2 flows

The unit tests at this stage are better for at least the eIDAS flow
and future work will ensure better/full coverage.