EID-634: Add jersey client names to example configs
The default client name is the same for the metadata resolver
and the trust anchor resolver which causes a clash in the metric
repository. Give it an explicitly different name in the example
configuration files to avoid this.
This can/should later be fixed in the library to use a different
default name in the eIDAS resolver but this will prevent the issue now