EID-866: Making MSA service end point support eidasEnabled and shouldSignWithSHA1
Enhancing the MatchingResponseGenerator so that it can create a Response that also contains
the two new flags.  This required enhancing the HealthCheckResponseFromMatchingService data
container class so that can create a request ID that now includes extra information, in this
format:

healthcheck-response-id-{0}-version-{1}-eidasenabled-{2}-shouldsignwithsha1-{3}

Note that changes were required in the hub so that it could parse eidasenabled and
shouldsignwithsha1 values.

Also note we're adding the semantics of field values into an identifier that is probably not
intended to have it.  We're using this approach as an alternative to introducing new SAML attributes
that would do the same thing.  We acknowledge this may not be great if many more variables are
added in this way but we anticipate in the short term this is a more practical solution.

Authors: @kgarwood, @felixmarcusmillne