EID-1016: Make "environment" in the MSA config a required value
This change ensures that the config clearly specifies the correct metadata truststores to be used by the MSA. Previously, it was possible to inadvertently configure the wrong truststores without any obvious message as the config defaults to INTEGRATION if environment is not specified.

This change ensures that either the 'environment' value is specified - in which case the default prod or integration truststores shipped with the MSA will be used, or that both 'idpTrustStore' and 'hubTrustStore' are set, in which case the RPs can use their own truststores.

Author: @CharlesIC