EID-1072 Pull in build 68 of saml-libs
It does three things.

Updates `EidasMetadataResolverRepository` to throw an Exception instead
of an Error if it encounters and expired cert in the trust anchor. This
will allow the MSA to startup, rather than crashing.

The MedataResolverBundle now sets up the metadata healthcheck, so the
one registered in the Application class can be removed.

Makes metadata bundling optional. This requires an update to the
configuration class, specifically `getMetadataConfiguration` to return
an Optional. The MSA config annotates the metadata with `@NotNull` so we
should never get a null optional, however `orElseThrow` has been added
in certain places with a new Exception to be on the safe side.