Accept multiple current attributes
In some circumstances, IDPs are sending results to the MSA with
multiple current attributes. These are usually due to, for example,
a name having multiple forms, one of which is verified.
This change ensures that multiple current calues does not cause an
error in the MSA - as these errors propagate to the Hub and fail user
journeys unnecessarily - but instead the verified value will be used
when available.

Note for future reference that if attempting to combine the two
comparator lines into one, the result of the intermediary call needs
to be explicitly typed to prevent Java "non-static method cannot be
referenced from a static context" errors.
(i.e. `Comparator.<SimpleMdsValue<T>, Boolean>comparing`)

Co-authored-by: Brendan Butler <brendan.butler@digital.cabinet-office.gov.uk>