Consolidate comparators
The parent commit added a comparator to the project that almost
duplicated another. The only difference is that simple MDS values
were needed instead of transliterable ones. Since the latter are a
superclass, the new comparator is sufficient for the use case of the
existing comparator. The existing comparator has been extracted to a
new file for partial reuse.

The reuse only needs to be partial because the original duplication
forgot to account for the filtering already being applied to the
data.