Add config files for MSA backward compatibility testing
This adds config files to run supported versions of the MSA for
backward compatibility testing.

In this commit: 3.1.0-840, the oldest supported version.

Also, a cloud foundry manifest template for running on PaaS.

I have hardcoded as much as possible specifically to the
backward-compatibility testing use case.  This means we only deploy to
staging so we don't need to template out entity ids or metadata URLs
or whatever.