HUB-697: Load metadata trust stores by default
The MSA was not using the trust stores for the federation metadata that
are supplied in the `resources` for the MSA. Instead it required users
to configure the path and password to a trust store in the `metadata`
section of their config file.

The current config files provided for prod and test had hard coded
values for these trust stores which pointed to a trust store in the root
of the project. These trust stores are actually only included when the
scripts to create an MSA release are run. They script [grabs particular
metadata certificates directly from the verify-pki-certs repo.](https://github.com/alphagov/verify-infrastructure-config/blob/master/pipelines/deployer/deploy-stubs.yml#L629)

This is opaque and also brittle. Without being aware of this it would be
easy to think you were adding a new generation of certificates to the
trust stores contained in the resources folder, and end users would
actually not be using them.

This commit allows the metadata trust store to have a default (the
resources trust store), in the same way we currently do for the hub and
IDP trust store. This means that users no longer need to define the
`truststore` in the `metadata` section of the config. It consolidates
the classes used for the hub and IDP default configurations and re-uses
it for all three types of trust store.

It also updates the passwords (hard to see as their a binary file) for
the metadata trust stores so the default password will work.

This will be a breaking change. If a user upgrades to this version of
the MSA but doesn't change their config file to allow the default the
application will fail to load as the previous trust stores that were
bundled by the deploy pipeline will be removed.