HUB-539: Keep Concourse MSA version in step with local version
The concourse build pipeline uses a semver resource to keep track of
different builds of the MSA. Every new PR that's merged triggers a build
and the pipeline bumps the patch version. This new version is used as
the version number for a draft github release.

When we want to generate a new release of the MSA for use by RP's we
bump the version in the `settings.gradle`. This will almost always
either include a major or minor bump.

This commit allows the build pipeline to pick up that change and update
the Concourse version to match. This will then produce a draft release
on Github with the correct version which we can then promote to a full
release.