HUB-697: Remove trust store config from LMS ci
As part of releasing the VSP we spin up an MSA: https://cd.gds-reliability.engineering/teams/verify/pipelines/verify-service-provider/jobs/deploy-msa-in-integration

This MSA is just pulled from master and uses the config that we're
changing here. That config is still looking for the old trust stores so
we need to remove it. There will also be a change to the pipeline to
stop providing the variables to the manifest.