BAU: Add hub trust store for back compat tests
The staging hub has moved to using 100 year signing and encryption
certs. These certs are from the dev-pki repo and have a different route
of trusts to the previous hub certs.

The old versions of the MSA came bundled with a trust store for the hub
which was correct at the time.

To keep the tests running, we need to drop in an updated trust store for
them to use. This means they'll be able to trust the new hub certs.

The pipeline will deal with putting this trust store in the right place.
The config is changed to look for it there.