HUB-812: Add dev-pki and G3 CAs to back compat tests
The versions of the MSA used for the backwards compatability tests don't
include trust stores which contain the dev-pki or G3 CAs. We want to use
these PKIs in staging and integration.

Version 5.0.2 of the MSA DOES have the CAs in its trust stores so
doesn't need to use these custom trust stores.

We have a service that we know is still using version 4.2.1 but with
upgraded custom trust stores, which is why we still need to support it
by doing this work.