HUB-912: Pull from maven central
Bintray and jcentre are going away. This updates the public builds to
pull from maven central instead. It also bumps the versions of our
libraries to ensure they're coming from maven central.

It also removes the ben maines plugin which was sold old it didn't work
anymore.

This also pins the source compatibility to Java 8. We're still
maintaining support for Java 8 for the services, and if features from
later versions are used we won't be able to compile for Java 8.

Also bump gradle to 6.8.2