BAU: Update version 5.2.3
This adds no functional changes. It's being bumped to align version in
the pipeline.

The Concourse pipeline maintains it's own MSA version semver resource
which is used when publishing releases to GitHub. The pipeline bumps the
patch version of this semver resource every time a change is merged.

This means that if you want to release a patch version to GitHub it's
very difficult - the patch number on the semver resource will already be
a few versions ahead of the version defined in `settings.gradle`.

We should decouple the GitHub release version from Concourse's internal
semver resources, but for now this should bring the two into alignment
to get the release out.